requirements.txt
opencv-python-headless
streamlit
numpy
Pillow
kociemba
scikit-learn
joblib
ultralytics
